+++ projects/all_data.py
@@ -1,0 +1,477 @@
+# There are 3 structures required for a quarter:
+# XXX_YEAR_PROJECT: This is a list of lists
+#   items should be:
+#        1. name
+#        2. url
+#        3. mentor
+#        4. TA
+#        5. github location
+#        6. flag 1/0 for if repo is private or not
+#        7. if there is a one-pager 1/0
+#        8. External Mentor Info
+#        9. if the URL (item 2) can pass the github link test
+
+# XXX_YEAR_STUDENT is a list of lists
+#   items should be:
+#        1. Project (matches to the project in PROJECT)
+#        2. Student Name (what appears on the webpage)
+#        3. github url OR just the github user name (if the text does not
+#                begins with https:// then "https://github.com/" will be added
+
+# XXX_YEAR_NAME_MAP
+# This is an override for the name. It replaces what is used internally with
+# something readable. It is a dictionary with the name link to the new name
+
+empty_23_PROJECT = []
+
+empty_23_STUDENT = []
+
+empty_23_NAME_MAP = {}
+
+WINTER_23_NAME_MAP = {"PCDC": "Pediatric Cancer Data Commons"}
+
+WINTER_23_PROJECT = [
+    [
+        "Argonne",
+        "https://www.anl.gov/",
+        "Rahim & Trevor",
+        None,
+        "https://github.com/dsi-clinic/2023-clinic-Argonne",
+        1,
+        1,
+        "[Matthew Dearing](https://scholar.google.com/citations?user=HUQIELDxZkgJ&hl=en)",
+        1,
+    ],
+    [
+        "BankTrack",
+        "https://www.banktrack.org/",
+        "Trevor",
+        None,
+        "https://github.com/chicago-cdac/banktrack-loan-pipeline/",
+        1,
+        1,
+        "XXX",
+        1,
+    ],
+    [
+        "Blue Ocean Gear",
+        "https://www.blueoceangear.com/",
+        "Trevor",
+        "Todd",
+        "https://github.com/chicago-cdac/bog-anomaly-mapping",
+        1,
+        1,
+        "XXX",
+        1,
+    ],
+    [
+        "BPI",
+        "https://www.impactforequity.org/",
+        "Riley",
+        "Kenia",
+        "https://github.com/dsi-clinic/2022-bpi-clinic",
+        1,
+        1,
+        "<ul><li>Amy Thompson</li><li>Loren Jones</li></ul>",
+        1,
+    ],
+    [
+        "DRW",
+        "https://drw.com/",
+        "Nick",
+        None,
+        "https://github.com/dsi-clinic/2023-clinic-drw",
+        1,
+        0,
+        "Ian Adam",
+        1,
+    ],
+    [
+        "Fermi",
+        "https://computing.fnal.gov/michael-kirby/",
+        "Peter",
+        "Todd",
+        "https://github.com/dsi-clinic/2023-clinic-fermi-tag",
+        1,
+        1,
+        "<ul><li>Michael Kirby</li><li>Meghna Bhattacharya</li></ul>",
+        1,
+    ],
+    [
+        "First Republic Bank",
+        "https://www.firstrepublic.com/",
+        "Nick",
+        None,
+        "https://github.com/dsi-clinic/2023-clinic-first-republic-bank",
+        1,
+        0,
+        '"<ul><li>Chris Csiszar</li><li>Mark Woodworth</li></ul>"',
+        1,
+    ],
+    [
+        "GreenWave",
+        "https://www.greenwave.org/",
+        "Trevor",
+        None,
+        "https://github.com/dsi-clinic/2023-clinic-greenwave",
+        1,
+        1,
+        "XXX",
+        1,
+    ],
+    [
+        "Invenergy",
+        "https://invenergy.com/",
+        "Anna",
+        "UT",
+        "https://github.com/dsi-clinic/2022-Invenergy-clinic",
+        1,
+        0,
+        "<ul><li>Zoë Kimpel</li><li>Kenneth Parkhill</li></ul>",
+        1,
+    ],
+    [
+        "Morningstar",
+        "https://www.morningstar.com/",
+        "David U. & Patricia",
+        "UT",
+        "https://github.com/dsi-clinic/2023-clinic-morningstar",
+        1,
+        0,
+        "<ul><li>Josh Charney</li><li>Jazmin Melchor</li></ul>",
+        1,
+    ],
+    [
+        "PCDC",
+        "https://commons.cri.uchicago.edu/pcdc/",
+        "Tim",
+        None,
+        "https://github.com/chicago-cdac/2023-clinic-pcdc",
+        1,
+        1,
+        "Michael Watkins",
+        1,
+    ],
+    [
+        "Perpetual",
+        "https://www.perpetualuse.org/",
+        "Rahim",
+        None,
+        "https://github.com/dsi-clinic/2023-clinic-perpetual",
+        0,
+        1,
+        "Ellie Moss",
+        1,
+    ],
+    [
+        "Prudential",
+        "https://www.prudential.com/",
+        "Nick",
+        None,
+        None,
+        1,
+        0,
+        "Amol Tembe",
+        1,
+    ],
+    [
+        "RISC",
+        "https://risc.uchicago.edu/",
+        "Jeffrey",
+        "Anthony",
+        "https://github.com/dsi-clinic/2023-clinic-risc",
+        1,
+        1,
+        "Noah Duncan",
+        1,
+    ],
+    [
+        "Neurocritical Care",
+        "https://profiles.uchicago.edu/profiles/display/17338286",
+        "Yuetian",
+        "Anthony",
+        "https://github.com/dsi-clinic/2023-clinic-neurocritical-care",
+        1,
+        1,
+        "Dr. Ali Mansour",
+        1,
+    ],
+    [
+        "Internet Equity",
+        "https://internetequity.uchicago.edu/",
+        "James Turk",
+        "Kenia",
+        "https://github.com/chicago-cdac/broadbandequity",
+        1,
+        1,
+        "Nicole Marwell",
+        1,
+    ],
+]
+
+WINTER_23_STUDENT = [
+    ["Argonne", "Soren Dunn", "https://github.com/sorendunn"],
+    ["Argonne", "Richard Huang", "https://github.com/rrhuang"],
+    ["Argonne", "Grace Shao", "https://github.com/graceshaoy"],
+    ["BankTrack", "Nivedita Vatsa", "https://github.com/nivedita-k-vatsa"],
+    ["BankTrack", "Yifu Hou", "https://github.com/yifu-hou"],
+    ["BankTrack", "Grishma Bhattarai", "https://github.com/grishmab"],
+    ["BankTrack", "Gillian Major", "https://github.com/gillianmajor"],
+    ["Blue Ocean Gear", "Gautam Kapoor", "https://github.com/grkapoor17"],
+    ["Blue Ocean Gear", "Ming-Chieh Liu", "https://github.com/ming-chieh-liu"],
+    ["Blue Ocean Gear", "Henry herzog", "https://github.com/hgherzog"],
+    ["BPI", "Ashley Hitchings", "https://github.com/ashleyhitchings"],
+    ["BPI", "Justin Kim", "https://github.com/jykim21"],
+    ["BPI", "Akila Forde", "https://github.com/aforde17"],
+    ["DRW", "Mahnoor Khan", "https://github.com/Mfk-han"],
+    ["DRW", "Xin Li", "https://github.com/xin2006"],
+    ["DRW", "Jasmeet Singh Sandhu", "https://github.com/jasmeeetSingh"],
+    ["DRW", "Yulun Han", "https://github.com/YLHan97"],
+    ["Fermi", "Richard Zhang", None],
+    ["Fermi", "Yuxin Ji", "https://github.com/Yuxin-Ji"],
+    ["Fermi", "Jason Zhang", "https://github.com/Zhang-QC"],
+    ["First Republic Bank", "Zhiyun Hu", "https://github.com/zhiyun0707"],
+    ["First Republic Bank", "Guangbo Niu", "https://github.com/ngbdsb"],
+    ["First Republic Bank", "Yu-Hsuan Chou", "https://github.com/yhchou0904"],
+    ["First Republic Bank", "Ning Tang", "https://github.com/tangn121"],
+    ["GreenWave", "Cole von Glahn", "https://github.com/cvg117"],
+    ["GreenWave", "Piper Kurtz", "https://github.com/kurtzpuc"],
+    ["GreenWave", "Nico Vila Alarcon", "https://github.com/niicovila"],
+    ["Invenergy", "David Bukowski", "https://github.com/dtbukowski"],
+    ["Invenergy", "Daisuke Kageyama", "https://github.com/daisukekageyama"],
+    ["Invenergy", "Baichen Tan", "https://github.com/BaichenTan"],
+    ["Invenergy", "Suyash Lakhani", "https://github.com/"],
+    ["Morningstar", "Emily Yeh", "https://github.com/Emily-fyeh"],
+    ["Morningstar", "Ruiquan Chang", "https://github.com/rqchang"],
+    [
+        "Morningstar",
+        "Pedro Antonio Ramonetti Vega",
+        "https://github.com/PRAMONETTI",
+    ],
+    ["Morningstar", "Nicole Li", "https://github.com/linicoley"],
+    ["PCDC", "Xuerong Shang", "https://github.com/xuerong98"],
+    ["PCDC", "Yu Zhou (Zoey)", "https://github.com/zoeyzhou1296"],
+    ["PCDC", "Shuhan Liu", "https://github.com/ShannaLiu"],
+    ["Perpetual", "Izzy Allum", "https://github.com/iallum"],
+    ["Perpetual", "Ziyu Ren", "https://github.com/AshleyZR"],
+    ["Perpetual", "Sandra Mauro", "https://github.com/sandramauro"],
+    ["Perpetual", "Yushu Qiu", "https://github.com/yushuqiu1"],
+    ["Prudential", "Yujing Sun", "https://github.com/yujing-syj"],
+    ["Prudential", "Cole Silva", "https://github.com/silva-cole"],
+    ["Prudential", "Nayna Pashilkar", "https://github.com/npashilkar"],
+    ["RISC", "Carolyn Liu", "https://github.com/Crliu4"],
+    ["RISC", "Avery Schoen", "https://github.com/averyschoen"],
+    ["RISC", "Adil Kassim", "https://github.com/adilkassim"],
+    [
+        "Neurocritical Care",
+        "Alexander Przybycin",
+        "https://github.com/AlexPrizzy",
+    ],
+    [
+        "Neurocritical Care",
+        "Oishee Chakrabarti",
+        "https://github.com/chakraoishee",
+    ],
+    [
+        "Neurocritical Care",
+        "Zachary Rothstein",
+        "https://github.com/Zacharyr41",
+    ],
+    ["Neurocritical Care", "Jim Tinley", "https://github.com/jtinley0"],
+    ["Internet Equity", "Maia Boyd", "https://github.com/maiaboyd"],
+    ["Internet Equity", "Victoria Kielb", "https://github.com/vkielb"],
+    ["Internet Equity", "Kaya Borlase", "https://github.com/borlasekn"],
+    ["Internet Equity", "Brendon Krall", "https://github.com/bkrall36"],
+]
+
+
+SPRING_23_PROJECT = [
+    [
+        "Argonne",
+        "https://www.anl.gov/",
+        "Rahim",
+        "Christian",
+        "https://github.com/dsi-clinic/2023-clinic-Argonne",
+        1,
+        1,
+        "[Matthew Dearing](https://scholar.google.com/citations?user=HUQIELDxZkgJ&hl=en)",
+        1,
+    ],
+    [
+        "Blue Ocean Gear",
+        "https://www.blueoceangear.com/",
+        "Launa",
+        None,
+        "https://github.com/chicago-cdac/bog-anomaly-mapping/",
+        1,
+        1,
+        "XX?",
+        1,
+    ],
+    [
+        "CRI-SET",
+        "https://pediatrics.uchicago.edu/research/set",
+        "Anna & Dan N.",
+        "Anthony K",
+        "https://github.com/dsi-clinic/2023-spring-clinic-set",
+        1,
+        1,
+        "Dr. Henry David",
+        1,
+    ],
+    [
+        "DRW",
+        "https://drw.com/",
+        "Tim",
+        None,
+        "https://github.com/dsi-clinic/2023-clinic-drw",
+        1,
+        0,
+        "Ian Adam",
+        1,
+    ],
+    [
+        "Fermi",
+        "https://computing.fnal.gov/michael-kirby/",
+        "Peter",
+        "Ali",
+        "https://github.com/dsi-clinic/2023-clinic-fermi-tag",
+        1,
+        1,
+        "<ul><li>Michael Kirby</li><li>Meghna Bhattacharya</li></ul>",
+        1,
+    ],
+    [
+        "FRB",
+        "https://www.firstrepublic.com/",
+        "Nick",
+        None,
+        "https://github.com/dsi-clinic/2023-clinic-first-republic-bank",
+        1,
+        0,
+        "<ul><li>Chris Csiszar</li><li>Mark Woodworth</li></ul>",
+        1,
+    ],
+    [
+        "Hawaii",
+        "https://tsffoundation.org/",
+        "Launa",
+        "Ali",
+        "https://github.com/chicago-cdac/hawaii-pesticides",
+        1,
+        1,
+        "XXX",
+        0,
+    ],
+    [
+        "IE",
+        "https://internetequity.uchicago.edu/",
+        "James Turk",
+        "Kenia",
+        "https://github.com/chicago-cdac/broadbandequity",
+        0,
+        1,
+        "Dr. Nicole Marwell",
+        1,
+    ],
+    [
+        "Morningstar",
+        "https://www.morningstar.com/",
+        "Patricia & David U.",
+        "Christian",
+        "https://github.com/dsi-clinic/2023-clinic-morningstar",
+        1,
+        0,
+        "<ul><li>Josh Charney</li><li>Jazmin Melchor</li></ul>",
+        1,
+    ],
+    [
+        "Neurocritical Care",
+        "https://scholar.google.com/citations?user=cs_tgvwAAAAJ&hl=en",
+        "Yuetian",
+        "Anthony",
+        "https://github.com/dsi-clinic/2023-clinic-neurocritical-care",
+        1,
+        1,
+        "Dr. Ali Mansour",
+        1,
+    ],
+    [
+        "Perpetual",
+        "https://www.perpetualuse.org/",
+        "Rahim",
+        None,
+        "https://github.com/dsi-clinic/2023-clinic-perpetual",
+        0,
+        1,
+        "Ellie Moss",
+        1,
+    ],
+    [
+        "Prudential",
+        "https://www.prudential.com/",
+        "Nick",
+        "Kenia",
+        None,
+        1,
+        0,
+        "Dave Powers",
+        1,
+    ],
+]
+
+SPRING_23_STUDENT = [
+    ["Argonne", "Ken Kliesner", "https://github.com/kenkliesner"],
+    ["Argonne", "Annabel Mendoza", "https://github.com/amendoza5025"],
+    ["Argonne", "Kekun Han", "https://github.com/KekunH"],
+    ["Blue Ocean Gear", "Gautam Kapoor", "https://github.com/grkapoor17"],
+    ["Blue Ocean Gear", "Henry Herzog", "https://github.com/Hgherzog"],
+    ["Blue Ocean Gear", "Irsa Ashraf", "https://github.com/irsa-ashraf"],
+    ["Blue Ocean Gear", "Katy Barone", "https://github.com/kbarone"],
+    ["CRI-SET", "Varun Mohan", "https://github.com/vmohan96"],
+    ["CRI-SET", "Jun Tan", "https://github.com/JunTan2022"],
+    ["CRI-SET", "Katherine Miao", "https://github.com/Katherine-Miao"],
+    ["DRW", "Mahnoor Khan", "https://github.com/Mfk-han"],
+    ["DRW", "Jasmeet Singh Sandhu", "https://github.com/jasmeeetSingh"],
+    ["DRW", "Yulun Han", "https://github.com/YLHan97"],
+    ["Fermi", "Richard Zhang", None],
+    ["Fermi", "Manuel Martinez", "https://github.com/manmartgarc"],
+    ["Fermi", "Mingyan Wang", "https://github.com/wmingyan"],
+    ["Fermi", "Tarun Arora", "https://github.com/tarun2k"],
+    ["FRB", "Guangbo Niu", "https://github.com/ngbdsb"],
+    ["FRB", "Zhiyun Hu", "https://github.com/zhiyun0707"],
+    ["FRB", "Yu-Hsuan Chou", "https://github.com/yhchou0904"],
+    ["FRB", "Ning Tang", "https://github.com/tangn121"],
+    ["Hawaii", "Ashley Hitchings", "https://github.com/ashleyhitchings"],
+    ["Hawaii", "Qingyi He", "https://github.com/cindyheqy"],
+    ["Hawaii", "Caleb Costa", "https://github.com/calebcosta1"],
+    ["IE", "Victoria Kielb", "https://github.com/vkielb"],
+    ["IE", "Chandler Hall", "https://github.com/cgwhall"],
+    ["IE", "Sarah Lueling", "https://github.com/slueling"],
+    ["Morningstar", "Rishabh Shastry", "https://github.com/rishabhshastry"],
+    ["Morningstar", "Dhairya Karna", "https://github.com/DhairyaKarna"],
+    ["Morningstar", "Max de Saint-Exupery", "https://github.com/MaxSaint01"],
+    ["Neurocritical Care", "Soren Dunn", "https://github.com/sorendunn"],
+    ["Neurocritical Care", "Alex Przybycin", "https://github.com/AlexPrizzy"],
+    [
+        "Neurocritical Care",
+        "Prashant Kumar",
+        "https://github.com/Prashant-Kumar700",
+    ],
+    ["Perpetual", "Ziyu Ren", "https://github.com/AshleyZR"],
+    ["Perpetual", "Yushu Qiu", "https://github.com/yushuqiu1"],
+    ["Perpetual", "Avery Schoen", "https://github.com/averyschoen"],
+    ["Perpetual", "Ekansh Trivedi", "https://github.com/ekanshtrivedi"],
+    ["Prudential", "Sunvid Aneja", "https://github.com/sunvidaneja"],
+    ["Prudential", "Peihan Gao", "https://github.com/peihan12"],
+    ["Prudential", "Sai Omkar Kandukuri", "https://github.com/S-Omkar-K"],
+    ["Prudential", "Hantang Qin", "https://github.com/jenniferqinnn"],
+]
+
+SPRING_23_NAME_MAP = {
+    "Argonne": "Argonne Knowledge Graph",
+    "Fermi": "Fermi: Kirby Lab",
+    "Hawaii": "Hawaii Pesticides",
+    "IE": "Internet Equity",
+    "FRB": "First Republic Bank",
+}
