+++ projects/gen_projects_md.py
@@ -1,0 +1,211 @@
+# Checklist:
+# 1. Remove access to named students when going through repos
+# 2. For repos in the dsi-clinic org: add the "all-nda" group to repos
+# 3. For complex 11th hour repos
+
+# To do:
+# 1. automate 400 no-400 for has repo
+
+from all_data import SPRING_23_NAME_MAP, SPRING_23_PROJECT, SPRING_23_STUDENT
+from all_data import WINTER_23_NAME_MAP, WINTER_23_PROJECT, WINTER_23_STUDENT
+
+PREAMBLE_TEXT = """### Previous Projects
+
+This page contains a list of projects organized by quarter with the list of students who worked on the project and the faculty mentor. Note that the dates listed below are by _calendar year_ not academic year.
+
+A few important notes:
+* Not all projects have complete information. Some of the projects are under NDAs or other specifications (such as UChicago not hosting the repo).
+* The requirements for the project have changed year over year and are sometimes project specific.
+
+---
+"""
+
+
+# Format: Name which links : [Name to display, github link] -- don't use LI as it breaks.
+ALL_PEOPLE = {
+    "James Turk": ["James Turk", "https://github.com/jamesturk/"],
+    "Nick": ["Nick Ross", "https://www.nickross.site/"],
+    "Launa": ["Launa Greer", "https://github.com/LaunaG"],
+    "Rahim": ["Rahim Rasool", "https://github.com/rahimrasool"],
+    "David U.": [
+        "David Uminsky",
+        "https://cs.uchicago.edu/people/david-uminsky/",
+    ],
+    "Dan N.": [
+        "Dan Nicolae",
+        "https://www.stat.uchicago.edu/~nicolae/",
+    ],
+    "Tim": ["Tim Hannifan", "https://github.com/timhannifan"],
+    "Ali": ["Ali Klemencic", "https://github.com/aliklemencic"],
+    "Christian": ["Christian Jordan", "https://github.com/chrjor"],
+    "Anthony K": ["Anthony Kanellopoulos", "https://github.com/kanello"],
+    "Kenia": ["Kenia Godinez Nogueda", "https://github.com/gnogueda"],
+    "Patricia": [
+        "Patricia Chiril",
+        "https://scholar.google.com/citations?user=AzsyeyIAAAAJ&hl=en",
+    ],
+    "Peter": ["Peter Lu", "https://github.com/peterparity"],
+    "Anna": [
+        "Anna Woodard",
+        "https://scholar.google.com/citations?user=1Gs8kcYAAAAJ&hl=en",
+    ],
+    "Todd": ["Todd Nief", "https://github.com/toddnief"],
+    "Yuetian": ["Yuetian Luo", "https://yuetianluo.github.io/"],
+    "Victor": ["Victor Perez Martin", "https://github.com/vperezmartin"],
+    "UT": ["Utkarsh Tripathi", "https://github.com/redgene"],
+    "Avery": ["Avery Schoen", "https://github.com/averyschoen"],
+    "Sunvid": ["Sunvid Aneja", "https://github.com/sunvidaneja"],
+    "Trevor" : ["Trevor Spreadbury", "https://github.com/trevorspreadbury"],
+    "Riley" : ["Riley Tucker", "https://scholar.google.com/citations?user=j8TVqU8AAAAJ&hl=en"],
+    "Jeffrey" : ["Jeffrey Negrea", "https://scholar.google.ca/citations?user=woSzLBMAAAAJ&hl=en"],
+}
+
+
+def create_link_for_mentor(mentor_info):
+    if mentor_info[1]:
+        return f"[{mentor_info[0]}]({mentor_info[1]})"
+    else:
+        return f"{mentor_info[0]}"
+
+
+def create_link_for_student(student_info):
+    # If the github username is only the username preprend the github url.
+    if student_info[2] == None or student_info[2][0:8] == "https://":
+        return create_link_for_mentor(student_info[1:3])
+    else:
+        return create_link_for_mentor(
+            [student_info[1], f"https://www.github.com/{student_info[2]}"]
+        )
+
+
+def create_single_quarter_table(
+    name_map,
+    project_map,
+    one_pager_location,
+    student_info_list,
+    quarter_name=None,  # noqa
+):
+    """
+    This returns a single table of information.
+    """
+
+    all_results = "\n| Project Name | Repository | One-Pager | Mentors | Students | \
+        External Mentor | TA | \n | --- |  --- | --- | --- | --- | --- | --- |\n"
+
+    for project_info in project_map:
+        # Loop over each project in project_map as the main loop
+        [
+            project_link,
+            project_url,
+            mentor_link,
+            ta_link,
+            github_link,
+            is_private_repo,
+            has_one_pager,
+            external_mentor_info,
+            project_url_valid,
+        ] = project_info
+
+        # Project name -- replace with what is in name map if it exists
+        project_name = name_map.get(project_link, project_link)
+
+        # Add url to the name. If the invalid flag is 1, add flag to not check
+        # link via the markdown link checker
+        if project_url_valid:
+            project_name_info = f"[{project_name}]({project_url})"
+        else:
+            project_name_info = f"<!-- markdown-link-check-disable -->[{project_name}]({project_url})<!-- markdown-link-check-enable -->"
+
+        # If the repo is private than the markdown link checker will fail, so
+        # add an exclusion for it in the case of a private repo
+        if is_private_repo:
+            if github_link:
+                repo_info = f"<!-- markdown-link-check-disable --> [Private Repo]({github_link}) <!-- markdown-link-check-enable -->"
+            else:
+                repo_info = "No Repository"
+        else:
+            repo_info = f"[DSI Repo]({github_link})"
+
+        # If there is a one pager than add a link to it.
+        if has_one_pager:
+            file_info = one_pager_location + project_link + ".pdf"
+            one_pager_info = f"[One-Pager]({file_info.replace(' ', '%20')})"
+        else:
+            one_pager_info = ""
+
+        # Mentor list. split by "&". If there are more than one then make a bullet point list
+        mentor_list = [mentor.strip() for mentor in mentor_link.split("&")]
+
+        if len(mentor_list) == 1:
+            mentor_info = create_link_for_mentor(ALL_PEOPLE[mentor_list[0]])
+
+        else:
+            mentor_info = "<ul>"
+            for mentor in mentor_list:
+                mentor_info += (
+                    f"<li>{create_link_for_mentor(ALL_PEOPLE[mentor])}</li>"
+                )
+            mentor_info += "</ul>"
+
+        # Student info. Assume that there is more than one and make a list.
+        student_info = "<ul>"
+        student_project_list = [x for x in student_info_list if x[0] == project_link]
+        if len(student_project_list) == 0:
+            raise Exception(f"No Students found for project {project_link}")
+        for student in student_project_list:
+            student_info += f"<li>{create_link_for_student(student)}</li>"
+        student_info += "</ul>"
+
+        # TA -- same logic as above, but handle the case with no TA
+        TA_info = ALL_PEOPLE.get(ta_link, None)
+        if TA_info:
+            TA_info = create_link_for_mentor(TA_info)
+        else:
+            TA_info = ""
+
+        project_line = " | ".join(
+            [
+                project_name_info,
+                repo_info,
+                one_pager_info,
+                mentor_info,
+                student_info,
+                external_mentor_info,
+                TA_info,
+            ]
+        )
+        all_results += " | " + project_line + "\n"
+
+    return all_results
+
+
+if __name__ == "__main__":
+    all_quarter_info_list = [
+        {
+            "quarter_name": "Spring 2023",
+            "name_map": SPRING_23_NAME_MAP,
+            "one_pager_location": "./one-pagers/2023-spring/",
+            "student_info_list": SPRING_23_STUDENT,
+            "project_map": SPRING_23_PROJECT,
+        },
+        {
+            "quarter_name": "Winter 2023",
+            "name_map": WINTER_23_NAME_MAP,
+            "one_pager_location": "./one-pagers/2023-winter/",
+            "student_info_list": WINTER_23_STUDENT,
+            "project_map": WINTER_23_PROJECT,
+        }
+
+    ]
+
+    with open("projects.md", "w") as f_handle:
+        f_handle.write(PREAMBLE_TEXT)
+        f_handle.write("\n")
+
+        for quarter in all_quarter_info_list:
+            f_handle.write("\n<details>\n\n")
+            f_handle.write(
+                f"<summary><h2>{quarter['quarter_name']}</h2></summary>\n\n"
+            )
+            f_handle.write(create_single_quarter_table(**quarter))
+            f_handle.write("\n</details>")
